requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pymupdf>=1.24.0
playwright>=1.40.0
anthropic>=0.18.0
streamlit>=1.32.0
pandas>=2.0.0
python-dotenv>=1.0.0
